Fix form data capture issue by implementing prefixes for multiple forms
Resolved a problem where form data was not being captured correctly due to the presence of
multiple forms with overlapping field names. Added unique prefixes to each form, ensuring
that data is properly bound and validated upon submission.

Updated code within the views

- Updated `BillingAddressForm` with prefix "billing_address_form".
- Updated `ShippingAddressForm` with prefix "shipping_address_form".
- Ensures unique identification of form fields when handling multiple forms on the same page,
  preventing data conflicts during submission.

The forms now behaves correctly by submitting as well as saving in the datat to the db