**Feat:** Add the ability to delete either shipping or billing address
- Added the functionality to delete either a shipping or billing address.
- Created a partial called `profile_sidebar` and added it to the partials.
- Modified the sidebar to include the following links: `Add Address`, `Manage Shipping Address`, and `Manage Billing Address`.
  Clicking these links allows users to view the corresponding addresses.

The function correctly deletes the address, whether it is shipping or billing, and enables users to add any address.

**TODO:**
- Add functionality to mark any address as primary within the UI.
- Add the ability for the shipping section to display "No more addresses" when there are no addresses available instead of just blank space. when
the page is refreshed it correctly shows the right message if there are no more `addresses` but this will also be done when the last address is removed
by the `fetch` request.