**feat:** Add a view that handles changing of passwords url
**Overview**
This commit introduces a new view and form to manage user password changes effectively, addressing both user
verification and password strength requirements.

**View Creation**
Created a view called `new_password` inside the authentication views folder that handles the forgotten password URL.
When the user clicks on the URL sent to their email, the function checks if the URL is valid, whether it belongs to the user,
and if it has expired. If all these conditions are met, the user can change their password. The view function also prevents
users from reusing the link if they have already changed their password.

**Form Creation**
Developed a new password form in `authentication/forms/passwords/new_password.py` that validates the password.
It checks for requirements such as length, presence of numbers, lowercase, uppercase, special characters, etc.
If these criteria are met, the user can change their password; otherwise, a form error message is displayed.

**Process**
The form allows users to request a new password from the login page, change their password,
and log in using the new password. It also displays appropriate error messages if any fields are invalid or if the token has been used.

**Todo**
1. Write tests for the new functionality.

2. Add a suppress mechanism that temporarily disables the button in the `forgotten password email form`` to prevent users from spamming the button
and hitting the views multiple times.

3. Refactor `generate_verification_url` and `generate_forgotten_password_url` functions to eliminate redundancy by
  extracting the shared logic into a single abstract function.

**Rationale for refactoring `generate_verification_url` and `generate_forgotten_password_url` **
The current implementation had significant overlap because there are duplicates within the code meaning it only differs in one or two place and the rest
are the same. I will apply the DRY principle to reduce the duplication in both functions by moving the shared code into an abstract function
which help improve the readability and reduces the risk of inconsistencies when changes are needed.