Implement testimonial submission process
When a user navigates to the account section, selects 'Manage Reviews', and fills in the testimonial
form under 'Add Testimonial', a notification message is sent to the admin, informing them that a new
testimonial is ready for review.

TODO: Once the admin approves the testimonial, an email notification
should be sent to the user to inform them of the approval status.