Feat: Add fetch to enable the user to be able to subscribe without refreshing the page
- Added a fetch inside `subscribeFormHandler.js` that allows the user to be able to subscribe to the sites
newsletter.

- Added a context_processor inside the subscription app folder so that this accessible anywhere. this
allows for a more personal greeting based on whether the user has subscribed. If the user has subscribed
they are greeted with a tailor message and if not their are greeted with a subscription form. The tailor
message is persistence even after refresh because it is located inside the sessions. The application also
checks that email the user is trying to log in is one that they have logged in with before, if it is
a message is display informing them of that

- Made the subscribe form only visible to login in users. This will allow users to be able to log into
their account and be able to set the frequency of the newsletter e.g weekly, monthly,etc or even be able
to unsubscribe. At the moment the UI interface for viewing the subscrbption and be able to manage the subscrbption
is not yet availalbe, that will be created in later

- Remove the following JS files that was powering that subscrbption form because they were now unneccessary now that back end was built
   - subscribeElements.js (This was used because the backend was yet built and was doing all the functionality mentioned in the above commit)

- Remove the jwtToken.js since a secret token is now implemented in the backend
   - jwtToken.js

Workflow
- The user logs into their account, goes to the homepage by clicking logo icon, scrolls down to the bottom of the page and subscribes.
- The user is then greeted with a tailor message