**Feature: Make the Reply Link Interactive on the `blog_details.html` Page**
The reply link on the blog details page has been made interactive. When clicked, a reply form appears,
allowing the user to respond directly to a specific comment. Clicking the reply link again will close the reply form.

This functionality has been implemented by adding a JavaScript file named `blog.js`, which contains the essential
logic to enable interactivity for the reply link. This script is also responsible for dynamically generating
the reply form and managing its visibility based on user interaction. The rationale for creating the form
dynamically is that it should only be instantiated when the corresponding reply link is clicked.

**Additional Changes:**

1. Rectified the broken image issue in the card for the CS.
2. Introduced an "OP" label for the author. Once the backend is integrated, any message posted by the
  Original Poster (OP) of the comment will be displayed in the card.
3. Reduced the size of the button by decreasing its padding.

**Workflow:**

1. No login is required.
2. Navigate to the blog section via the navigation toolbar.
3. Click on any card to access the blog details page.
4. Clicking on any of the reply links located beneath a comment will now reveal the reply form.