**Feat**: Add functionality to allow a user to like a post
Implemented functionality enabling users to like a post, including the necessary CSS, HTML, and
JavaScript. Upon navigating to the `blogs_detail.html` page, users can click the `thumbs up`
icon to like the post. Currently, users can like the post multiple times, incrementing the like counter.
However, once the backend feature is implemented, users will only be able to like the post once and they
must be logged in to do so.
Most of the JavaScript was added to `blog.js`, but a `distanceUtils.js` file
 was also created inside the `js` utils folder to assist with the like functionality.

1. No login is required.
2. Navigate to the blog section via the navigation bar.
3. Click on any card to go to the blog's detail page.
4. Scroll to the section displaying the thumbs-up icon and click it.
5. The like counter will increase upon clicking.