**Feature: Fix for Subscription Error After Update**
In the previous version, users could subscribe via the form on the `index.html` page, and an email would be sent to the admin,
triggered by signals. This worked perfectly until I made changes to the subscription model but neglected to update the corresponding models in
`views.py`. As a result, the fields were not aligned with the updated model, which caused an error, preventing new subscribers from being registered.

**Fix:**
The issue was resolved by updating the fields in `views.py` to match those in the updated model.

**Additional Changes:**
- Added a link in the `index.html` page for users who have already subscribed and then unsubscribed, allowing them to go directly
  to the manage subscription section.
- Removed the 'Subscribe' button from the `manage-subscription.html` page if the user hasn’t subscribed yet, and if they navigate to
that page they see nothing. However, if the user has subscribed but later unsubscribed they will be show a resubscribe button.

- If the user subscribes from the `manage-subscription.html` page they are show a `re-subscribe` button

- Added a 'Re-subscribe' button for users who have unsubscribed. This allows them to re-subscribe easily.

**To Do:**
- Add a spinner to the `subscribe` button on the `index.html` page so users can see the progress when they click it.