**Feat: Add the ability to edit post**
1. Added the ability to `edit` a `reply` or a `comment` by clicking the "..." which is the access point for the control
panel located on every reply and comment. The control panel when clicked shows two fields.

    - edit
    - delete

2. When the user clicks the toggle control panel an editing box appears which allow
the user to edit their `comment` or a `reply` comment.  So, far the control panel access point "..." appears on every
comment, however when the backend functionality is created, the "..." will only appear on comments or reply that was
created by the user .

**Todo***

- Add an autoscroll to control the view, so that when the control panel is shown, it automatically scrolls into view.

Currently when the user clicks the "..." the conrol panel is diplayed. However, if the browser height is positioned
at a certain height, and the control panel access point is then clicked, the control box might not be shown and this
might depend on where the page view height is.