Feat: Add product details page
Added the initial step for a product details page. This page will display detailed information for any given product,
including colour and size options, description, dimensions, remaining stock, and more. Introduced the first stages of HTML and CSS,
with further refinements to follow for a visual aesthetically pleasing design.

Added the view and the url neccessary to make it work and display
- Added a nutrient json field to the model

Note:
Currently, the page can be accessed via "http://127.0.0.1:8000/product" for testing purposes and to see how it looks and behaviours. Once completed,
the URL will include an ID at the end via a slash, corresponding to each product.