Feat: Implement modular StarRating class for dynamic and interactive ratings by consolidating the functions in `renderStar.js`
- Added a `StarRating` class inside `renderStar.js` to handle the rendering and interaction of star ratings.
- Designed to be modular, enabling the class to be reused for rendering and interacting with star ratings on any component, as
   long as the correct HTML structure are provided.

- Configurable options include:
  - `numOfRatedStars`: Total number of stars to rated.
  - `filledStarsSrc` and `unfilledStarsSrc`: Paths to images for filled and unfilled stars.
  - `renderEmptyStars`: Option to render unfilled stars even if no rating is selected.
  - `isInteractive`: Enables interactive mode where users can click to set ratings.
  - `totalNumOfStars`:The total number of unrated stars that will be used

- Provides methods to:
  - Initialise the star rating system with `initialise()`.
  - Dynamically update the star display using `_updateStarDisplay()`.
  - Clean up resources and event listeners with `destroy()`.

- Deleted the unecessary functions in `renderStar.js`

- Supports both static (read-only) and interactive use cases, making it versatile for different rating scenarios. This means it can display
- Integrated it to various js using `renderStar`

- HTML
- Added a few restriction the submit and ratings are only shown when the user logs in

Now when the page is navigated to and the user logs in, the user is able to click the stars which allows them to rate the product.
So, far it doesn't store the result or submit the form because it is not wired to the views

**Todo**
Make the colour and sizes clickable and selectable when a user clicks on it

**Known Issue**
When a user visits `http://127.0.0.1:8000/product` and clicks "Login," an error occurs due to redirection to `/authentication/login/`.

This happens for two reasons:

1. The view is designed to allow users to access the product details page without logging in. As a result, the necessary decorator that redirects
unauthenticated users to the homepage and adds a `next` URL is not included.

2. The login lacks a standalone page and can only be accessed via a dropdown menu on the homepage. This dropdown uses a `fetch`
request to log in via the `/authentication/login/` JSON endpoint. Attempting to log in from the product page incorrectly directs to this JSON endpoint,
causing an error.

**Current Workaround**
Users must log in through the homepage.