Feat: Add the fifth stage to be able to save to the database
Initial implementation of multi-step form navigation logic

Added the `redirect_to_incomplete_step` function to manage navigation for a multi-step form. The function prevents users
from proceeding to the next step if the current step is not completed, but at the same time allows them to revisit any previously completed step.

Includes:
- Logic to check completion status of each step.
- Redirection to the first incomplete step if accessed out of order.
- Support for rendering the current template if all conditions are met.
- modified the handle_form in the views and views_helper to except the current step. Each step represents a stage of multi-form step e.g first page = 1, second page = 2.... last_page = n

Refactor
 - Also created two new functions `validate_required_keys` and `validate_instance_of` to prevent the duplication and used them in `create_shipping_variations` and `create_product_variations`

This ensures that the user can't simply skip to the review page and then submit the page.

Todo
- Create the means to save to the database when the `submit` button is hit in the review page.